[generator] Move the type manager to its own file.
