+++ src/Makefile.generator
@@ -6,6 +6,7 @@ GENERATOR_SOURCES =                                             \
 	$(TOP)/src/generator-attributes.cs                 \
 	$(TOP)/src/generator-enums.cs                      \
 	$(TOP)/src/generator-filters.cs                    \
+	$(TOP)/src/generator-typemanager.cs                \
 	$(TOP)/src/ObjCRuntime/Stret.cs                    \
 	$(TOP)/tools/common/TargetFramework.cs             \
 	$(MONO_PATH)/mcs/class/Mono.Options/Mono.Options/Options.cs \
